Сохранить пакет controller.

+++ src/main/java/ru/yandex/practicum/filmorate/controller/ReviewController.java
@@ -30,40 +30,40 @@ public class ReviewController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable("id") Integer id) {
+    public void delete(@PathVariable("id") long id) {
         reviewService.delete(id);
     }
 
     @GetMapping
     public List<Review> getAll(
-            @RequestParam(name = "filmId", defaultValue = "-1", required = false) Integer filmId,
-            @RequestParam(name = "count", defaultValue = "10", required = false) Integer count
+            @RequestParam(name = "filmId", defaultValue = "-1", required = false) long filmId,
+            @RequestParam(name = "count", defaultValue = "10", required = false) int count
     ) {
         return reviewService.getAll(filmId, count);
     }
 
     @GetMapping("/{id}")
-    public Review getById(@PathVariable("id") Integer id) {
+    public Review getById(@PathVariable("id") long id) {
         return reviewService.getById(id);
     }
 
     @PutMapping("/{id}/like/{userId}")
-    public void addUserLike(@PathVariable("id") Integer id, @PathVariable("userId") Integer userId) {
+    public void addUserLike(@PathVariable("id") long id, @PathVariable("userId") long userId) {
         reviewService.addLikeToFilmReview(id, userId);
     }
 
     @PutMapping("/{id}/dislike/{userId}")
-    public void addUserDislike(@PathVariable("id") Integer id, @PathVariable("userId") Integer userId) {
+    public void addUserDislike(@PathVariable("id") long id, @PathVariable("userId") long userId) {
         reviewService.addDislikeToFilmReview(id, userId);
     }
 
     @DeleteMapping("/{id}/like/{userId}")
-    public void deleteUserLike(@PathVariable("id") Integer id, @PathVariable("userId") Integer userId) {
+    public void deleteUserLike(@PathVariable("id") long id, @PathVariable("userId") long userId) {
         reviewService.deleteLikeFromFilmReview(id, userId);
     }
 
     @DeleteMapping("/{id}/dislike/{userId}")
-    public void deleteUserDislike(@PathVariable("id") Integer id, @PathVariable("userId") Integer userId) {
+    public void deleteUserDislike(@PathVariable("id") long id, @PathVariable("userId") long userId) {
         reviewService.deleteDislikeFromFilmReview(id, userId);
     }
 }
